storage_read.c: clarify anti-corrupted-cache-data check
This could have been part of
    2015-06-27 Don't add corrupt data to the block cache.
    fc7675b806cde989a8acde76595293ef764c3d52

That bugfix made a minimal change to the previous code by adding
    && (sc == 0)
to the previous "should we cache this data?" check.  However, this
resulted in a long if() check, wherein 3 of the checks are about the
cache, and 1 additional check is about the earlier code in this function.

I think the code is more clear if we ask "did we decrypt the data
successfully?  Ok, now go do a bunch of cache-related stuff".